+++ projects/ausmittlung-lib/src/lib/models/proportional-election-result.model.ts
@@ -39,6 +39,7 @@ export {
   ProportionalElectionResultBundlesProto,
   ProportionalElectionResultBundleProto,
   ProportionalElectionResultBallotProto,
+  ProportionalElectionListResultProto,
   ProportionalElectionListResultsProto,
   ProportionalElectionResultSubTotalProto,
   ProportionalElectionListResultSubTotalProto,