+++ projects/erfassung/src/app/app.component.ts
@@ -21,6 +21,7 @@ import { LocationStrategy } from '@angular/common';
 import { Title } from '@angular/platform-browser';
 import { filter } from 'rxjs/operators';
 import { RuntimeConfigService, LanguageService } from 'ausmittlung-lib';
+import { environment } from '../environments/environment';
 
 @Component({
   selector: 'app-root',
@@ -44,6 +45,7 @@ export class AppComponent implements OnInit, OnDestroy {
   public theme?: string;
   public customLogo?: string;
   public appTitle: string = '';
+  public customHeaderColor?: string;
 
   @ViewChild('snackbar')
   public snackbarComponent?: SnackbarComponent;
@@ -58,6 +60,8 @@ export class AppComponent implements OnInit, OnDestroy {
 
     // enable automatic silent refresh
     this.oauthService.setupAutomaticSilentRefresh({}, 'access_token');
+
+    this.customHeaderColor = environment.customHeaderColor;
 
     const snackbarSubscription = this.snackbarService.message$.subscribe(m => {
       if (!this.snackbarComponent) {