+++ projects/monitoring/src/environments/environment.model.ts
@@ -18,4 +18,5 @@ export interface Environment extends TenantConfig, UserConfig, AuthorizationConf
   includeDelegations: boolean;
   runtimeConfigPolling: RuntimeConfigPollingConfig;
   eventLogConfig: EventLogConfig;
+  customHeaderColor?: string;
 }